+++ CHANGES.rst
@@ -19,6 +19,7 @@ Major Changes
 Minor Changes
 ^^^^^^^^^^^^^
 
+- fix(concurrency): use timing-safe token comparisons & mask repr
 - fix(concurrency): rely on context manager boolean control flow
 - docs(coc): fix typos & rename first section
 - docs(perm): add missing operator in doc string examples